refactor(test): use directly `src/`
