--- index.js
@@ -1,7 +1,0 @@
-// LICENSE : MIT
-"use strict";
-module.exports = {
-    Store: require("./lib/material-store"),
-    Action: require("./lib/material-action"),
-    Context: require("./lib/material-context")
-};
