+++ src/index.js
@@ -1,0 +1,7 @@
+// LICENSE : MIT
+"use strict";
+module.exports = {
+    Store: require("./material-store"),
+    Action: require("./material-action"),
+    Context: require("./material-context")
+};
